+++ test/util/ReportTest.ts
@@ -1,0 +1,6 @@
+import { PremiseReport, SatisfactionState } from "../../src/util/report/ComplianceReportTypes";
+
+
+export function countSatisfiedPremises(premiseReports: PremiseReport[]): number {
+    return premiseReports.filter(premiseReport => premiseReport.satisfactionState === SatisfactionState.Satisfied).length
+}